pyproject.toml
[project]
name = "adk-crash-course"
version = "0.1.0"
description = "ADK Crash Course"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "certifi>=2025.4.26",
    "google-adk>=1.1.1",
    "google-generativeai>=0.8.5",
    "litellm>=1.72.0",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.0",
    "yfinance>=0.2.61",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
